requirements.txt
# Biology RAG System Requirements
chromadb==0.4.18
sentence-transformers==2.2.2
torch==2.1.0
requests==2.31.0
numpy==1.24.3
transformers==4.35.0
huggingface-hub==0.17.3
pinecone-client==6.0.0
tqdm==4.67.1
